# daemon/api_gen.go
//go:build go1.22

// Package daemon provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package daemon

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// Defines values for TrackCodec.
const (
	TrackCodecAac     TrackCodec = "aac"
	TrackCodecEmpty   TrackCodec = ""
	TrackCodecFlac    TrackCodec = "flac"
	TrackCodecMp3     TrackCodec = "mp3"
	TrackCodecUnknown TrackCodec = "unknown"
	TrackCodecVorbis  TrackCodec = "vorbis"
)

// ApiAddToQueue An add to queue payload
type ApiAddToQueue struct {
	// Uri The URI for the track that should be added
	Uri string `json:"uri"`
}

// ApiNext A skip to next payload
type ApiNext struct {
	// Uri The track URI to skip to. When omitted the next track in the context is played.
	Uri *string `json:"uri,omitempty"`
}

// ApiOutput A reopen audio output payload
type ApiOutput struct {
	// Device The output device name to open (backend-specific, e.g. an ALSA device). Empty selects the configured default.
	Device string `json:"device,omitempty"`
}

// ApiPlay An initiate playback payload
type ApiPlay struct {
	// Paused Start playback as paused
	Paused bool `json:"paused,omitempty"`

	// Position Position in milliseconds to start playback at within the selected track (0 starts from the beginning)
	Position int64 `json:"position,omitempty"`

	// SkipToUri Spotify URI to skip to (when playing playlists)
	SkipToUri string `json:"skip_to_uri,omitempty"`

	// Uri Spotify URI to start playing
	Uri string `json:"uri"`
}

// ApiRepeatContext A toggle repeating context payload
type ApiRepeatContext struct {
	// RepeatContext Whether repeating context should be enabled
	RepeatContext bool `json:"repeat_context"`
}

// ApiRepeatTrack A toggle repeating track payload
type ApiRepeatTrack struct {
	// RepeatTrack Whether repeating track should be enabled
	RepeatTrack bool `json:"repeat_track"`
}

// ApiRoot API reachability and playback readiness
type ApiRoot struct {
	// PlaybackReady Whether the daemon is fully bootstrapped and ready to accept /player/play
	PlaybackReady bool `json:"playback_ready"`
}

// ApiSeek A seek payload
type ApiSeek struct {
	// Position Seek position in milliseconds. Must not be negative unless relative is set.
	Position int64 `json:"position"`

	// Relative Whether the seek position is relative to the current one
	Relative bool `json:"relative,omitempty"`
}

// ApiSetDeviceName A set device name payload
type ApiSetDeviceName struct {
	// Name The new device name
	Name string `json:"name"`
}

// ApiSetVolume A set volume payload
type ApiSetVolume struct {
	// Relative Whether to change the volume relative to the current volume
	Relative bool `json:"relative,omitempty"`

	// Volume Volume from 0 to max. Must not be negative unless relative is set.
	Volume int32 `json:"volume"`
}

// ApiShuffleContext A toggle shuffling context payload
type ApiShuffleContext struct {
	// ShuffleContext Whether shuffling context should be enabled
	ShuffleContext bool `json:"shuffle_context"`
}

// ApiStatus The player status
type ApiStatus struct {
	// Buffering Whether the player is buffering
	Buffering bool `json:"buffering"`

	// DeviceId The player device ID
	DeviceId string `json:"device_id"`

	// DeviceName The player device name
	DeviceName string `json:"device_name"`

	// DeviceType The player device type, for example COMPUTER or SPEAKER
	DeviceType string `json:"device_type"`

	// Paused Whether the player is paused
	Paused bool `json:"paused"`

	// PlayOrigin Who started the playback, "go-librespot" identifies the API as the play origin, everything else is Spotify own stuff
	PlayOrigin string `json:"play_origin"`

	// RepeatContext Whether the repeat context feature is enabled
	RepeatContext bool `json:"repeat_context"`

	// RepeatTrack Whether the repeat track feature is enabled
	RepeatTrack bool `json:"repeat_track"`

	// ShuffleContext Whether the shuffle context feature is enabled
	ShuffleContext bool `json:"shuffle_context"`

	// Stopped Whether the player is stopped
	Stopped bool      `json:"stopped"`
	Track   *ApiTrack `json:"track"`

	// Username Currently active account's username
	Username string `json:"username"`

	// Volume The player current volume from 0 to max
	Volume uint32 `json:"volume"`

	// VolumeSteps The player max volume value
	VolumeSteps uint32 `json:"volume_steps"`
}

// ApiToken A Spotify access token for the active session
type ApiToken struct {
	// Token The access token
	Token string `json:"token"`
}

// ApiTrack A track or podcast episode
type ApiTrack struct {
	// AlbumCoverUrl Album cover URL
	AlbumCoverUrl *string `json:"album_cover_url"`

	// AlbumName Album name
	AlbumName string `json:"album_name"`

	// ArtistNames Artists name
	ArtistNames []string `json:"artist_names"`

	// BitDepth Bits per sample of the source audio, null for lossy formats which have no meaningful source bit depth
	BitDepth *int `json:"bit_depth"`

	// Bitrate Nominal bitrate in kbps, null for formats without a fixed one such as FLAC
	Bitrate *int `json:"bitrate"`

	// Codec Codec family of the audio file
	Codec TrackCodec `json:"codec"`

	// DiscNumber Disc number within the album, zero for episodes
	DiscNumber int `json:"disc_number"`

	// Duration Duration in milliseconds
	Duration int `json:"duration"`

	// Format Spotify name of the audio file being decoded, for example OGG_VORBIS_160
	Format string `json:"format"`

	// Name Name
	Name string `json:"name"`

	// Position Playback position in milliseconds
	Position int64 `json:"position"`

	// ReleaseDate Album release date, empty for episodes
	ReleaseDate string `json:"release_date"`

	// SampleRate Sample rate of the decoded audio in Hz
	SampleRate *int `json:"sample_rate"`

	// TrackNumber Track number within the disc, zero for episodes
	TrackNumber int `json:"track_number"`

	// Uri URI
	Uri string `json:"uri"`
}

// TrackCodec Codec family of the audio file
type TrackCodec string

// ApiVolume The player volume settings
type ApiVolume struct {
	// Max The max volume value
	Max uint32 `json:"max"`

	// Value The current volume, ranging from 0 to max
	Value uint32 `json:"value"`
}

// PlayerStreamParams defines parameters for PlayerStream.
type PlayerStreamParams struct {
	// Uri Spotify track or episode URI, for example spotify:track:4iV5W9uYEdYUVa79Axb7Rh
	Uri string `form:"uri" json:"uri"`

	// Bitrate Preferred bitrate in kbps. When omitted the daemon's configured bitrate is used.
	Bitrate *int `form:"bitrate,omitempty" json:"bitrate,omitempty"`
}

// PlayerAddToQueueJSONRequestBody defines body for PlayerAddToQueue for application/json ContentType.
type PlayerAddToQueueJSONRequestBody = ApiAddToQueue

// PlayerNextJSONRequestBody defines body for PlayerNext for application/json ContentType.
type PlayerNextJSONRequestBody = ApiNext

// PlayerOutputJSONRequestBody defines body for PlayerOutput for application/json ContentType.
type PlayerOutputJSONRequestBody = ApiOutput

// PlayerPlayJSONRequestBody defines body for PlayerPlay for application/json ContentType.
type PlayerPlayJSONRequestBody = ApiPlay

// PlayerRepeatContextJSONRequestBody defines body for PlayerRepeatContext for application/json ContentType.
type PlayerRepeatContextJSONRequestBody = ApiRepeatContext

// PlayerRepeatTrackJSONRequestBody defines body for PlayerRepeatTrack for application/json ContentType.
type PlayerRepeatTrackJSONRequestBody = ApiRepeatTrack

// PlayerSeekJSONRequestBody defines body for PlayerSeek for application/json ContentType.
type PlayerSeekJSONRequestBody = ApiSeek

// PlayerShuffleContextJSONRequestBody defines body for PlayerShuffleContext for application/json ContentType.
type PlayerShuffleContextJSONRequestBody = ApiShuffleContext

// PlayerSetVolumeJSONRequestBody defines body for PlayerSetVolume for application/json ContentType.
type PlayerSetVolumeJSONRequestBody = ApiSetVolume

// SetDeviceNameJSONRequestBody defines body for SetDeviceName for application/json ContentType.
type SetDeviceNameJSONRequestBody = ApiSetDeviceName

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /)
	GetRoot(w http.ResponseWriter, r *http.Request)

	// (GET /events)
	GetEvents(w http.ResponseWriter, r *http.Request)

	// (POST /player/add_to_queue)
	PlayerAddToQueue(w http.ResponseWriter, r *http.Request)

	// (POST /player/next)
	PlayerNext(w http.ResponseWriter, r *http.Request)

	// (POST /player/output)
	PlayerOutput(w http.ResponseWriter, r *http.Request)

	// (POST /player/pause)
	PlayerPause(w http.ResponseWriter, r *http.Request)

	// (POST /player/play)
	PlayerPlay(w http.ResponseWriter, r *http.Request)

	// (POST /player/playpause)
	PlayerPlayPause(w http.ResponseWriter, r *http.Request)

	// (POST /player/prev)
	PlayerPrev(w http.ResponseWriter, r *http.Request)

	// (POST /player/repeat_context)
	PlayerRepeatContext(w http.ResponseWriter, r *http.Request)

	// (POST /player/repeat_track)
	PlayerRepeatTrack(w http.ResponseWriter, r *http.Request)

	// (POST /player/resume)
	PlayerResume(w http.ResponseWriter, r *http.Request)

	// (POST /player/seek)
	PlayerSeek(w http.ResponseWriter, r *http.Request)

	// (POST /player/shuffle_context)
	PlayerShuffleContext(w http.ResponseWriter, r *http.Request)

	// (POST /player/stop)
	PlayerStop(w http.ResponseWriter, r *http.Request)

	// (GET /player/stream)
	PlayerStream(w http.ResponseWriter, r *http.Request, params PlayerStreamParams)

	// (GET /player/volume)
	PlayerGetVolume(w http.ResponseWriter, r *http.Request)

	// (POST /player/volume)
	PlayerSetVolume(w http.ResponseWriter, r *http.Request)

	// (POST /set_device_name)
	SetDeviceName(w http.ResponseWriter, r *http.Request)

	// (GET /status)
	GetStatus(w http.ResponseWriter, r *http.Request)

	// (POST /token)
	GetToken(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetRoot operation middleware
func (siw *ServerInterfaceWrapper) GetRoot(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRoot(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetEvents operation middleware
func (siw *ServerInterfaceWrapper) GetEvents(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetEvents(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerAddToQueue operation middleware
func (siw *ServerInterfaceWrapper) PlayerAddToQueue(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerAddToQueue(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerNext operation middleware
func (siw *ServerInterfaceWrapper) PlayerNext(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerNext(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerOutput operation middleware
func (siw *ServerInterfaceWrapper) PlayerOutput(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerOutput(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerPause operation middleware
func (siw *ServerInterfaceWrapper) PlayerPause(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerPause(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerPlay operation middleware
func (siw *ServerInterfaceWrapper) PlayerPlay(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerPlay(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerPlayPause operation middleware
func (siw *ServerInterfaceWrapper) PlayerPlayPause(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerPlayPause(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerPrev operation middleware
func (siw *ServerInterfaceWrapper) PlayerPrev(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerPrev(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerRepeatContext operation middleware
func (siw *ServerInterfaceWrapper) PlayerRepeatContext(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerRepeatContext(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerRepeatTrack operation middleware
func (siw *ServerInterfaceWrapper) PlayerRepeatTrack(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerRepeatTrack(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerResume operation middleware
func (siw *ServerInterfaceWrapper) PlayerResume(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerResume(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerSeek operation middleware
func (siw *ServerInterfaceWrapper) PlayerSeek(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerSeek(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerShuffleContext operation middleware
func (siw *ServerInterfaceWrapper) PlayerShuffleContext(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerShuffleContext(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerStop operation middleware
func (siw *ServerInterfaceWrapper) PlayerStop(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerStop(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerStream operation middleware
func (siw *ServerInterfaceWrapper) PlayerStream(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params PlayerStreamParams

	// ------------- Required query parameter "uri" -------------

	if paramValue := r.URL.Query().Get("uri"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "uri"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "uri", r.URL.Query(), &params.Uri)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "uri", Err: err})
		return
	}

	// ------------- Optional query parameter "bitrate" -------------

	err = runtime.BindQueryParameter("form", true, false, "bitrate", r.URL.Query(), &params.Bitrate)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bitrate", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerStream(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerGetVolume operation middleware
func (siw *ServerInterfaceWrapper) PlayerGetVolume(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerGetVolume(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PlayerSetVolume operation middleware
func (siw *ServerInterfaceWrapper) PlayerSetVolume(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PlayerSetVolume(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SetDeviceName operation middleware
func (siw *ServerInterfaceWrapper) SetDeviceName(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SetDeviceName(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStatus operation middleware
func (siw *ServerInterfaceWrapper) GetStatus(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStatus(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetToken operation middleware
func (siw *ServerInterfaceWrapper) GetToken(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetToken(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/{$}", wrapper.GetRoot)
	m.HandleFunc("GET "+options.BaseURL+"/events", wrapper.GetEvents)
	m.HandleFunc("POST "+options.BaseURL+"/player/add_to_queue", wrapper.PlayerAddToQueue)
	m.HandleFunc("POST "+options.BaseURL+"/player/next", wrapper.PlayerNext)
	m.HandleFunc("POST "+options.BaseURL+"/player/output", wrapper.PlayerOutput)
	m.HandleFunc("POST "+options.BaseURL+"/player/pause", wrapper.PlayerPause)
	m.HandleFunc("POST "+options.BaseURL+"/player/play", wrapper.PlayerPlay)
	m.HandleFunc("POST "+options.BaseURL+"/player/playpause", wrapper.PlayerPlayPause)
	m.HandleFunc("POST "+options.BaseURL+"/player/prev", wrapper.PlayerPrev)
	m.HandleFunc("POST "+options.BaseURL+"/player/repeat_context", wrapper.PlayerRepeatContext)
	m.HandleFunc("POST "+options.BaseURL+"/player/repeat_track", wrapper.PlayerRepeatTrack)
	m.HandleFunc("POST "+options.BaseURL+"/player/resume", wrapper.PlayerResume)
	m.HandleFunc("POST "+options.BaseURL+"/player/seek", wrapper.PlayerSeek)
	m.HandleFunc("POST "+options.BaseURL+"/player/shuffle_context", wrapper.PlayerShuffleContext)
	m.HandleFunc("POST "+options.BaseURL+"/player/stop", wrapper.PlayerStop)
	m.HandleFunc("GET "+options.BaseURL+"/player/stream", wrapper.PlayerStream)
	m.HandleFunc("GET "+options.BaseURL+"/player/volume", wrapper.PlayerGetVolume)
	m.HandleFunc("POST "+options.BaseURL+"/player/volume", wrapper.PlayerSetVolume)
	m.HandleFunc("POST "+options.BaseURL+"/set_device_name", wrapper.SetDeviceName)
	m.HandleFunc("GET "+options.BaseURL+"/status", wrapper.GetStatus)
	m.HandleFunc("POST "+options.BaseURL+"/token", wrapper.GetToken)

	return m
}